fixed Equivalent Class Axiom with no Genus
Equivalent Class Axiom with no Genus

Problem: An equivalent class axiom of the kind C = R some A is nearly always an indication of a problem. It basically means that anything that is R related to an A is equal to the class. While this is sometimes legit, most often this is accidental.

Solution: Add a genus to the class expression like: C = B and R some A.